+++ R/ggpimage.R
@@ -20,10 +20,14 @@
 ## the rows and columns are swapped and the order of the
 ## columns (original rows) is reversed.
 
+#' @rdname pimage
+#' @include pimage.R
+#' @export
 ggpimage <- function(x,
   order = NULL,
   upper_tri = TRUE,
   lower_tri = TRUE,
+  diag = TRUE,
   row_labels = NULL,
   col_labels = NULL,
   prop = TRUE,
@@ -32,10 +36,13 @@ ggpimage <- function(x,
   UseMethod("ggpimage")
 
 ### Note for matrix large values are dark, for dist large values are light!
+#' @rdname pimage
+#' @export
 ggpimage.matrix <- function(x,
   order = NULL,
   upper_tri = TRUE,
   lower_tri = TRUE,
+  diag = TRUE,
   row_labels = NULL,
   col_labels = NULL,
   prop = TRUE,
@@ -58,13 +65,16 @@ ggpimage.matrix <- function(x,
 
   # mask triangles
   if (any(!upper_tri ||
-      !lower_tri) &&
+      !lower_tri ||
+      !diag) &&
       nrow(x) != ncol(x))
-    stop("Upper or lower triangle can only be suppressed for square matrices!")
+    stop("Upper triangle, lower triangle or diag can only be suppressed for square matrices!")
   if (!upper_tri)
     x[upper.tri(x)] <- NA
   if (!lower_tri)
     x[lower.tri(x)] <- NA
+  if (!diag)
+    diag(x) <- NA
 
 
   # reverse order of columns
@@ -224,11 +234,14 @@ ggpimage.matrix <- function(x,
 ggpimage.default <- ggpimage.matrix
 
 ## small values are dark
+#' @rdname pimage
+#' @export
 ggpimage.dist <-
   function(x,
     order = NULL,
     upper_tri = FALSE,
     lower_tri = TRUE,
+    diag = FALSE,
     row_labels = NULL,
     col_labels = NULL,
     prop = TRUE,
@@ -248,6 +261,7 @@ ggpimage.dist <-
       order = NULL,
       upper_tri,
       lower_tri,
+      diag,
       row_labels,
       col_labels,
       prop = prop,
@@ -257,8 +271,7 @@ ggpimage.dist <-
 
     # reverse color for dist
     suppressMessages(g <-
-        g + .gg_sequential_pal(dist = TRUE)
-    )
+        g + .gg_sequential_pal(dist = TRUE))
 
     g
   }